Add scoring when asteroid is destroyed
When a bullet collides with an asteroid, destroy the asteroid and
bullet and increase the score by 1.

So when we have lots of overlapping asteroid and bullet bounding
boxes, we only want to count the an asteroid as collideded with the
first bullet encountered and should not be counted again.